% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_functions.R
\name{acl.insert}
\alias{acl.insert}
\title{Creates an access control rule.}
\usage{
acl.insert(AclRule, calendarId)
}
\arguments{
\item{AclRule}{The \link{AclRule} object to pass to this method}

\item{calendarId}{Calendar identifier}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/calendar
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/calendar)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/calendar/firstapp}{Google Documentation}

Other AclRule functions: \code{\link{AclRule.scope}},
  \code{\link{AclRule}}, \code{\link{acl.patch}},
  \code{\link{acl.update}}
}
