% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{Acl}
\alias{Acl}
\title{Calendar API Objects 
Manipulates events and other calendar data.}
\usage{
Acl(etag = NULL, items = NULL, nextPageToken = NULL,
  nextSyncToken = NULL)
}
\arguments{
\item{etag}{ETag of the collection}

\item{items}{List of rules on the access control list}

\item{nextPageToken}{Token used to access the next page of this result}

\item{nextSyncToken}{Token used at a later point in time to retrieve only the entries that have changed since this result was returned}
}
\value{
Acl object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-04 00:00:52
filename: /Users/mark/dev/R/autoGoogleAPI/googlecalendarv3.auto/R/calendar_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Acl Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

