% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{AclRule}
\alias{AclRule}
\title{AclRule Object}
\usage{
AclRule(AclRule.scope = NULL, etag = NULL, id = NULL, role = NULL,
  scope = NULL)
}
\arguments{
\item{AclRule.scope}{The \link{AclRule.scope} object or list of objects}

\item{etag}{ETag of the resource}

\item{id}{Identifier of the ACL rule}

\item{role}{The role assigned to the scope}

\item{scope}{The scope of the rule}
}
\value{
AclRule object
}
\description{
AclRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other AclRule functions: \code{\link{AclRule.scope}},
  \code{\link{acl.insert}}, \code{\link{acl.patch}},
  \code{\link{acl.update}}
}

