% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{AclRule.scope}
\alias{AclRule.scope}
\title{AclRule.scope Object}
\usage{
AclRule.scope(type = NULL, value = NULL)
}
\arguments{
\item{type}{The type of the scope}

\item{value}{The email address of a user or group, or the name of a domain, depending on the scope type}
}
\value{
AclRule.scope object
}
\description{
AclRule.scope Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The scope of the rule.
}
\seealso{
Other AclRule functions: \code{\link{AclRule}},
  \code{\link{acl.insert}}, \code{\link{acl.patch}},
  \code{\link{acl.update}}
}
