% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{Calendar}
\alias{Calendar}
\title{Calendar Object}
\usage{
Calendar(description = NULL, etag = NULL, id = NULL, location = NULL,
  summary = NULL, timeZone = NULL)
}
\arguments{
\item{description}{Description of the calendar}

\item{etag}{ETag of the resource}

\item{id}{Identifier of the calendar}

\item{location}{Geographic location of the calendar as free-form text}

\item{summary}{Title of the calendar}

\item{timeZone}{The time zone of the calendar}
}
\value{
Calendar object
}
\description{
Calendar Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Calendar functions: \code{\link{.insert}},
  \code{\link{.patch}}, \code{\link{.update}}
}
