% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_functions.R
\docType{package}
\name{calendar_googleAuthR}
\alias{calendar_googleAuthR}
\alias{calendar_googleAuthR-package}
\title{Calendar API
Manipulates events and other calendar data.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:30:35
filename: /Users/mark/dev/R/autoGoogleAPI/googlecalendarv3.auto/R/calendar_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/calendar
\item https://www.googleapis.com/auth/calendar.readonly
}
}
