% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{CalendarList}
\alias{CalendarList}
\title{CalendarList Object}
\usage{
CalendarList(etag = NULL, items = NULL, nextPageToken = NULL,
  nextSyncToken = NULL)
}
\arguments{
\item{etag}{ETag of the collection}

\item{items}{Calendars that are present on the user's calendar list}

\item{nextPageToken}{Token used to access the next page of this result}

\item{nextSyncToken}{Token used at a later point in time to retrieve only the entries that have changed since this result was returned}
}
\value{
CalendarList object
}
\description{
CalendarList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
