% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{CalendarListEntry.notificationSettings}
\alias{CalendarListEntry.notificationSettings}
\title{CalendarListEntry.notificationSettings Object}
\usage{
CalendarListEntry.notificationSettings(notifications = NULL)
}
\arguments{
\item{notifications}{The list of notifications set for this calendar}
}
\value{
CalendarListEntry.notificationSettings object
}
\description{
CalendarListEntry.notificationSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The notifications that the authenticated user is receiving for this calendar.
}
\seealso{
Other CalendarListEntry functions: \code{\link{CalendarListEntry}},
  \code{\link{ist.insert}}, \code{\link{ist.patch}},
  \code{\link{ist.update}}
}

