% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{CalendarListEntry}
\alias{CalendarListEntry}
\title{CalendarListEntry Object}
\usage{
CalendarListEntry(CalendarListEntry.notificationSettings = NULL,
  accessRole = NULL, backgroundColor = NULL, colorId = NULL,
  defaultReminders = NULL, deleted = NULL, description = NULL,
  etag = NULL, foregroundColor = NULL, hidden = NULL, id = NULL,
  location = NULL, notificationSettings = NULL, primary = NULL,
  selected = NULL, summary = NULL, summaryOverride = NULL,
  timeZone = NULL)
}
\arguments{
\item{CalendarListEntry.notificationSettings}{The \link{CalendarListEntry.notificationSettings} object or list of objects}

\item{accessRole}{The effective access role that the authenticated user has on the calendar}

\item{backgroundColor}{The main color of the calendar in the hexadecimal format '#0088aa'}

\item{colorId}{The color of the calendar}

\item{defaultReminders}{The default reminders that the authenticated user has for this calendar}

\item{deleted}{Whether this calendar list entry has been deleted from the calendar list}

\item{description}{Description of the calendar}

\item{etag}{ETag of the resource}

\item{foregroundColor}{The foreground color of the calendar in the hexadecimal format '#ffffff'}

\item{hidden}{Whether the calendar has been hidden from the list}

\item{id}{Identifier of the calendar}

\item{location}{Geographic location of the calendar as free-form text}

\item{notificationSettings}{The notifications that the authenticated user is receiving for this calendar}

\item{primary}{Whether the calendar is the primary calendar of the authenticated user}

\item{selected}{Whether the calendar content shows up in the calendar UI}

\item{summary}{Title of the calendar}

\item{summaryOverride}{The summary that the authenticated user has set for this calendar}

\item{timeZone}{The time zone of the calendar}
}
\value{
CalendarListEntry object
}
\description{
CalendarListEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other CalendarListEntry functions: \code{\link{CalendarListEntry.notificationSettings}},
  \code{\link{ist.insert}}, \code{\link{ist.patch}},
  \code{\link{ist.update}}
}
