% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{CalendarNotification}
\alias{CalendarNotification}
\title{CalendarNotification Object}
\usage{
CalendarNotification(method = NULL, type = NULL)
}
\arguments{
\item{method}{The method used to deliver the notification}

\item{type}{The type of notification}
}
\value{
CalendarNotification object
}
\description{
CalendarNotification Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
