% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{ColorDefinition}
\alias{ColorDefinition}
\title{ColorDefinition Object}
\usage{
ColorDefinition(background = NULL, foreground = NULL)
}
\arguments{
\item{background}{The background color associated with this color definition}

\item{foreground}{The foreground color that can be used to write on top of a background with 'background' color}
}
\value{
ColorDefinition object
}
\description{
ColorDefinition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
