% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{Colors.calendar}
\alias{Colors.calendar}
\title{Colors.calendar Object}
\usage{
Colors.calendar()
}
\value{
Colors.calendar object
}
\description{
Colors.calendar Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A global palette of calendar colors, mapping from the color ID to its definition. A calendarListEntry resource refers to one of these color IDs in its color field. Read-only.
}
\seealso{
Other Colors functions: \code{\link{Colors.event}},
  \code{\link{Colors}}
}
