% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{Colors.event}
\alias{Colors.event}
\title{Colors.event Object}
\usage{
Colors.event()
}
\value{
Colors.event object
}
\description{
Colors.event Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A global palette of event colors, mapping from the color ID to its definition. An event resource may refer to one of these color IDs in its color field. Read-only.
}
\seealso{
Other Colors functions: \code{\link{Colors.calendar}},
  \code{\link{Colors}}
}
