% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{Colors}
\alias{Colors}
\title{Colors Object}
\usage{
Colors(Colors.calendar = NULL, Colors.event = NULL, calendar = NULL,
  event = NULL, updated = NULL)
}
\arguments{
\item{Colors.calendar}{The \link{Colors.calendar} object or list of objects}

\item{Colors.event}{The \link{Colors.event} object or list of objects}

\item{calendar}{A global palette of calendar colors, mapping from the color ID to its definition}

\item{event}{A global palette of event colors, mapping from the color ID to its definition}

\item{updated}{Last modification time of the color palette (as a RFC3339 timestamp)}
}
\value{
Colors object
}
\description{
Colors Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Colors functions: \code{\link{Colors.calendar}},
  \code{\link{Colors.event}}
}

