% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{Error}
\alias{Error}
\title{Error Object}
\usage{
Error(domain = NULL, reason = NULL)
}
\arguments{
\item{domain}{Domain, or broad category, of the error}

\item{reason}{Specific reason for the error}
}
\value{
Error object
}
\description{
Error Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

