% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{Event.creator}
\alias{Event.creator}
\title{Event.creator Object}
\usage{
Event.creator(displayName = NULL, email = NULL, id = NULL, self = NULL)
}
\arguments{
\item{displayName}{The creator's name, if available}

\item{email}{The creator's email address, if available}

\item{id}{The creator's Profile ID, if available}

\item{self}{Whether the creator corresponds to the calendar on which this copy of the event appears}
}
\value{
Event.creator object
}
\description{
Event.creator Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The creator of the event. Read-only.
}
\seealso{
Other Event functions: \code{\link{Event.extendedProperties.private}},
  \code{\link{Event.extendedProperties.shared}},
  \code{\link{Event.extendedProperties}},
  \code{\link{Event.gadget.preferences}},
  \code{\link{Event.gadget}},
  \code{\link{Event.organizer}},
  \code{\link{Event.reminders}},
  \code{\link{Event.source}}, \code{\link{Event}},
  \code{\link{events.import}}, \code{\link{events.insert}},
  \code{\link{events.patch}}, \code{\link{events.update}}
}

