% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{Event.extendedProperties}
\alias{Event.extendedProperties}
\title{Event.extendedProperties Object}
\usage{
Event.extendedProperties(Event.extendedProperties.private = NULL,
  Event.extendedProperties.shared = NULL, private = NULL, shared = NULL)
}
\arguments{
\item{Event.extendedProperties.private}{The \link{Event.extendedProperties.private} object or list of objects}

\item{Event.extendedProperties.shared}{The \link{Event.extendedProperties.shared} object or list of objects}

\item{private}{Properties that are private to the copy of the event that appears on this calendar}

\item{shared}{Properties that are shared between copies of the event on other attendees' calendars}
}
\value{
Event.extendedProperties object
}
\description{
Event.extendedProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Extended properties of the event.
}
\seealso{
Other Event functions: \code{\link{Event.creator}},
  \code{\link{Event.extendedProperties.private}},
  \code{\link{Event.extendedProperties.shared}},
  \code{\link{Event.gadget.preferences}},
  \code{\link{Event.gadget}},
  \code{\link{Event.organizer}},
  \code{\link{Event.reminders}},
  \code{\link{Event.source}}, \code{\link{Event}},
  \code{\link{events.import}}, \code{\link{events.insert}},
  \code{\link{events.patch}}, \code{\link{events.update}}
}

