% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{Event.gadget.preferences}
\alias{Event.gadget.preferences}
\title{Event.gadget.preferences Object}
\usage{
Event.gadget.preferences()
}
\value{
Event.gadget.preferences object
}
\description{
Event.gadget.preferences Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Preferences.
}
\seealso{
Other Event functions: \code{\link{Event.creator}},
  \code{\link{Event.extendedProperties.private}},
  \code{\link{Event.extendedProperties.shared}},
  \code{\link{Event.extendedProperties}},
  \code{\link{Event.gadget}},
  \code{\link{Event.organizer}},
  \code{\link{Event.reminders}},
  \code{\link{Event.source}}, \code{\link{Event}},
  \code{\link{events.import}}, \code{\link{events.insert}},
  \code{\link{events.patch}}, \code{\link{events.update}}
}

