% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{Event.gadget}
\alias{Event.gadget}
\title{Event.gadget Object}
\usage{
Event.gadget(Event.gadget.preferences = NULL, display = NULL,
  height = NULL, iconLink = NULL, link = NULL, preferences = NULL,
  title = NULL, type = NULL, width = NULL)
}
\arguments{
\item{Event.gadget.preferences}{The \link{Event.gadget.preferences} object or list of objects}

\item{display}{The gadget's display mode}

\item{height}{The gadget's height in pixels}

\item{iconLink}{The gadget's icon URL}

\item{link}{The gadget's URL}

\item{preferences}{Preferences}

\item{title}{The gadget's title}

\item{type}{The gadget's type}

\item{width}{The gadget's width in pixels}
}
\value{
Event.gadget object
}
\description{
Event.gadget Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A gadget that extends this event.
}
\seealso{
Other Event functions: \code{\link{Event.creator}},
  \code{\link{Event.extendedProperties.private}},
  \code{\link{Event.extendedProperties.shared}},
  \code{\link{Event.extendedProperties}},
  \code{\link{Event.gadget.preferences}},
  \code{\link{Event.organizer}},
  \code{\link{Event.reminders}},
  \code{\link{Event.source}}, \code{\link{Event}},
  \code{\link{events.import}}, \code{\link{events.insert}},
  \code{\link{events.patch}}, \code{\link{events.update}}
}
