% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{Event.organizer}
\alias{Event.organizer}
\title{Event.organizer Object}
\usage{
Event.organizer(displayName = NULL, email = NULL, id = NULL,
  self = NULL)
}
\arguments{
\item{displayName}{The organizer's name, if available}

\item{email}{The organizer's email address, if available}

\item{id}{The organizer's Profile ID, if available}

\item{self}{Whether the organizer corresponds to the calendar on which this copy of the event appears}
}
\value{
Event.organizer object
}
\description{
Event.organizer Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The organizer of the event. If the organizer is also an attendee, this is indicated with a separate entry in attendees with the organizer field set to True. To change the organizer, use the move operation. Read-only, except when importing an event.
}
\seealso{
Other Event functions: \code{\link{Event.creator}},
  \code{\link{Event.extendedProperties.private}},
  \code{\link{Event.extendedProperties.shared}},
  \code{\link{Event.extendedProperties}},
  \code{\link{Event.gadget.preferences}},
  \code{\link{Event.gadget}},
  \code{\link{Event.reminders}},
  \code{\link{Event.source}}, \code{\link{Event}},
  \code{\link{events.import}}, \code{\link{events.insert}},
  \code{\link{events.patch}}, \code{\link{events.update}}
}
