% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{Event}
\alias{Event}
\title{Event Object}
\usage{
Event(Event.creator = NULL, Event.extendedProperties = NULL,
  Event.extendedProperties.private = NULL,
  Event.extendedProperties.shared = NULL, Event.gadget = NULL,
  Event.gadget.preferences = NULL, Event.organizer = NULL,
  Event.reminders = NULL, Event.source = NULL, anyoneCanAddSelf = NULL,
  attachments = NULL, attendees = NULL, attendeesOmitted = NULL,
  colorId = NULL, created = NULL, creator = NULL, description = NULL,
  end = NULL, endTimeUnspecified = NULL, etag = NULL,
  extendedProperties = NULL, gadget = NULL, guestsCanInviteOthers = NULL,
  guestsCanModify = NULL, guestsCanSeeOtherGuests = NULL,
  hangoutLink = NULL, htmlLink = NULL, iCalUID = NULL, id = NULL,
  location = NULL, locked = NULL, organizer = NULL,
  originalStartTime = NULL, privateCopy = NULL, recurrence = NULL,
  recurringEventId = NULL, reminders = NULL, sequence = NULL,
  source = NULL, start = NULL, status = NULL, summary = NULL,
  transparency = NULL, updated = NULL, visibility = NULL)
}
\arguments{
\item{Event.creator}{The \link{Event.creator} object or list of objects}

\item{Event.extendedProperties}{The \link{Event.extendedProperties} object or list of objects}

\item{Event.extendedProperties.private}{The \link{Event.extendedProperties.private} object or list of objects}

\item{Event.extendedProperties.shared}{The \link{Event.extendedProperties.shared} object or list of objects}

\item{Event.gadget}{The \link{Event.gadget} object or list of objects}

\item{Event.gadget.preferences}{The \link{Event.gadget.preferences} object or list of objects}

\item{Event.organizer}{The \link{Event.organizer} object or list of objects}

\item{Event.reminders}{The \link{Event.reminders} object or list of objects}

\item{Event.source}{The \link{Event.source} object or list of objects}

\item{anyoneCanAddSelf}{Whether anyone can invite themselves to the event (currently works for Google+ events only)}

\item{attachments}{File attachments for the event}

\item{attendees}{The attendees of the event}

\item{attendeesOmitted}{Whether attendees may have been omitted from the event's representation}

\item{colorId}{The color of the event}

\item{created}{Creation time of the event (as a RFC3339 timestamp)}

\item{creator}{The creator of the event}

\item{description}{Description of the event}

\item{end}{The (exclusive) end time of the event}

\item{endTimeUnspecified}{Whether the end time is actually unspecified}

\item{etag}{ETag of the resource}

\item{extendedProperties}{Extended properties of the event}

\item{gadget}{A gadget that extends this event}

\item{guestsCanInviteOthers}{Whether attendees other than the organizer can invite others to the event}

\item{guestsCanModify}{Whether attendees other than the organizer can modify the event}

\item{guestsCanSeeOtherGuests}{Whether attendees other than the organizer can see who the event's attendees are}

\item{hangoutLink}{An absolute link to the Google+ hangout associated with this event}

\item{htmlLink}{An absolute link to this event in the Google Calendar Web UI}

\item{iCalUID}{Event unique identifier as defined in RFC5545}

\item{id}{Opaque identifier of the event}

\item{location}{Geographic location of the event as free-form text}

\item{locked}{Whether this is a locked event copy where no changes can be made to the main event fields 'summary', 'description', 'location', 'start', 'end' or 'recurrence'}

\item{organizer}{The organizer of the event}

\item{originalStartTime}{For an instance of a recurring event, this is the time at which this event would start according to the recurrence data in the recurring event identified by recurringEventId}

\item{privateCopy}{Whether this is a private event copy where changes are not shared with other copies on other calendars}

\item{recurrence}{List of RRULE, EXRULE, RDATE and EXDATE lines for a recurring event, as specified in RFC5545}

\item{recurringEventId}{For an instance of a recurring event, this is the id of the recurring event to which this instance belongs}

\item{reminders}{Information about the event's reminders for the authenticated user}

\item{sequence}{Sequence number as per iCalendar}

\item{source}{Source from which the event was created}

\item{start}{The (inclusive) start time of the event}

\item{status}{Status of the event}

\item{summary}{Title of the event}

\item{transparency}{Whether the event blocks time on the calendar}

\item{updated}{Last modification time of the event (as a RFC3339 timestamp)}

\item{visibility}{Visibility of the event}
}
\value{
Event object
}
\description{
Event Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Event functions: \code{\link{Event.creator}},
  \code{\link{Event.extendedProperties.private}},
  \code{\link{Event.extendedProperties.shared}},
  \code{\link{Event.extendedProperties}},
  \code{\link{Event.gadget.preferences}},
  \code{\link{Event.gadget}},
  \code{\link{Event.organizer}},
  \code{\link{Event.reminders}},
  \code{\link{Event.source}}, \code{\link{events.import}},
  \code{\link{events.insert}}, \code{\link{events.patch}},
  \code{\link{events.update}}
}

