% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{Event.source}
\alias{Event.source}
\title{Event.source Object}
\usage{
Event.source(title = NULL, url = NULL)
}
\arguments{
\item{title}{Title of the source; for example a title of a web page or an email subject}

\item{url}{URL of the source pointing to a resource}
}
\value{
Event.source object
}
\description{
Event.source Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Source from which the event was created. For example, a web page, an email message or any document identifiable by an URL with HTTP or HTTPS scheme. Can only be seen or modified by the creator of the event.
}
\seealso{
Other Event functions: \code{\link{Event.creator}},
  \code{\link{Event.extendedProperties.private}},
  \code{\link{Event.extendedProperties.shared}},
  \code{\link{Event.extendedProperties}},
  \code{\link{Event.gadget.preferences}},
  \code{\link{Event.gadget}},
  \code{\link{Event.organizer}},
  \code{\link{Event.reminders}}, \code{\link{Event}},
  \code{\link{events.import}}, \code{\link{events.insert}},
  \code{\link{events.patch}}, \code{\link{events.update}}
}

