% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{EventAttachment}
\alias{EventAttachment}
\title{EventAttachment Object}
\usage{
EventAttachment(fileId = NULL, fileUrl = NULL, iconLink = NULL,
  mimeType = NULL, title = NULL)
}
\arguments{
\item{fileId}{ID of the attached file}

\item{fileUrl}{URL link to the attachment}

\item{iconLink}{URL link to the attachment's icon}

\item{mimeType}{Internet media type (MIME type) of the attachment}

\item{title}{Attachment title}
}
\value{
EventAttachment object
}
\description{
EventAttachment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
