% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{EventAttendee}
\alias{EventAttendee}
\title{EventAttendee Object}
\usage{
EventAttendee(additionalGuests = NULL, comment = NULL, displayName = NULL,
  email = NULL, id = NULL, optional = NULL, organizer = NULL,
  resource = NULL, responseStatus = NULL, self = NULL)
}
\arguments{
\item{additionalGuests}{Number of additional guests}

\item{comment}{The attendee's response comment}

\item{displayName}{The attendee's name, if available}

\item{email}{The attendee's email address, if available}

\item{id}{The attendee's Profile ID, if available}

\item{optional}{Whether this is an optional attendee}

\item{organizer}{Whether the attendee is the organizer of the event}

\item{resource}{Whether the attendee is a resource}

\item{responseStatus}{The attendee's response status}

\item{self}{Whether this entry represents the calendar on which this copy of the event appears}
}
\value{
EventAttendee object
}
\description{
EventAttendee Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

