% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{EventDateTime}
\alias{EventDateTime}
\title{EventDateTime Object}
\usage{
EventDateTime(date = NULL, dateTime = NULL, timeZone = NULL)
}
\arguments{
\item{date}{The date, in the format 'yyyy-mm-dd', if this is an all-day event}

\item{dateTime}{The time, as a combined date-time value (formatted according to RFC3339)}

\item{timeZone}{The time zone in which the time is specified}
}
\value{
EventDateTime object
}
\description{
EventDateTime Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
