% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{EventReminder}
\alias{EventReminder}
\title{EventReminder Object}
\usage{
EventReminder(method = NULL, minutes = NULL)
}
\arguments{
\item{method}{The method used by this reminder}

\item{minutes}{Number of minutes before the start of the event when the reminder should trigger}
}
\value{
EventReminder object
}
\description{
EventReminder Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

