% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_functions.R
\name{events.delete}
\alias{events.delete}
\title{Deletes an event.}
\usage{
events.delete(calendarId, eventId, sendNotifications = NULL)
}
\arguments{
\item{calendarId}{Calendar identifier}

\item{eventId}{Event identifier}

\item{sendNotifications}{Whether to send notifications about the deletion of the event}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/calendar
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/calendar)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/calendar/firstapp}{Google Documentation}
}

