% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_functions.R
\name{events.instances}
\alias{events.instances}
\title{Returns instances of the specified recurring event.}
\usage{
events.instances(calendarId, eventId, alwaysIncludeEmail = NULL,
  maxAttendees = NULL, maxResults = NULL, originalStart = NULL,
  pageToken = NULL, showDeleted = NULL, timeMax = NULL, timeMin = NULL,
  timeZone = NULL)
}
\arguments{
\item{calendarId}{Calendar identifier}

\item{eventId}{Recurring event identifier}

\item{alwaysIncludeEmail}{Whether to always include a value in the email field for the organizer, creator and attendees, even if no real email is available (i}

\item{maxAttendees}{The maximum number of attendees to include in the response}

\item{maxResults}{Maximum number of events returned on one result page}

\item{originalStart}{The original start time of the instance in the result}

\item{pageToken}{Token specifying which result page to return}

\item{showDeleted}{Whether to include deleted events (with status equals 'cancelled') in the result}

\item{timeMax}{Upper bound (exclusive) for an event's start time to filter by}

\item{timeMin}{Lower bound (inclusive) for an event's end time to filter by}

\item{timeZone}{Time zone used in the response}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/calendar
\item https://www.googleapis.com/auth/calendar.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/calendar, https://www.googleapis.com/auth/calendar.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/calendar/firstapp}{Google Documentation}
}

