% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_functions.R
\name{events.move}
\alias{events.move}
\title{Moves an event to another calendar, i.e. changes an event's organizer.}
\usage{
events.move(calendarId, eventId, destination, sendNotifications = NULL)
}
\arguments{
\item{calendarId}{Calendar identifier of the source calendar where the event currently is on}

\item{eventId}{Event identifier}

\item{destination}{Calendar identifier of the target calendar where the event is to be moved to}

\item{sendNotifications}{Whether to send notifications about the change of the event's organizer}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/calendar
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/calendar)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/calendar/firstapp}{Google Documentation}
}
