% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{Events}
\alias{Events}
\title{Events Object}
\usage{
Events(accessRole = NULL, defaultReminders = NULL, description = NULL,
  etag = NULL, items = NULL, nextPageToken = NULL, nextSyncToken = NULL,
  summary = NULL, timeZone = NULL, updated = NULL)
}
\arguments{
\item{accessRole}{The user's access role for this calendar}

\item{defaultReminders}{The default reminders on the calendar for the authenticated user}

\item{description}{Description of the calendar}

\item{etag}{ETag of the collection}

\item{items}{List of events on the calendar}

\item{nextPageToken}{Token used to access the next page of this result}

\item{nextSyncToken}{Token used at a later point in time to retrieve only the entries that have changed since this result was returned}

\item{summary}{Title of the calendar}

\item{timeZone}{The time zone of the calendar}

\item{updated}{Last modification time of the calendar (as a RFC3339 timestamp)}
}
\value{
Events object
}
\description{
Events Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
