% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{FreeBusyCalendar}
\alias{FreeBusyCalendar}
\title{FreeBusyCalendar Object}
\usage{
FreeBusyCalendar(busy = NULL, errors = NULL)
}
\arguments{
\item{busy}{List of time ranges during which this calendar should be regarded as busy}

\item{errors}{Optional error(s) (if computation for the calendar failed)}
}
\value{
FreeBusyCalendar object
}
\description{
FreeBusyCalendar Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

