% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{FreeBusyGroup}
\alias{FreeBusyGroup}
\title{FreeBusyGroup Object}
\usage{
FreeBusyGroup(calendars = NULL, errors = NULL)
}
\arguments{
\item{calendars}{List of calendars' identifiers within a group}

\item{errors}{Optional error(s) (if computation for the group failed)}
}
\value{
FreeBusyGroup object
}
\description{
FreeBusyGroup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
