% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{FreeBusyRequest}
\alias{FreeBusyRequest}
\title{FreeBusyRequest Object}
\usage{
FreeBusyRequest(calendarExpansionMax = NULL, groupExpansionMax = NULL,
  items = NULL, timeMax = NULL, timeMin = NULL, timeZone = NULL)
}
\arguments{
\item{calendarExpansionMax}{Maximal number of calendars for which FreeBusy information is to be provided}

\item{groupExpansionMax}{Maximal number of calendar identifiers to be provided for a single group}

\item{items}{List of calendars and/or groups to query}

\item{timeMax}{The end of the interval for the query}

\item{timeMin}{The start of the interval for the query}

\item{timeZone}{Time zone used in the response}
}
\value{
FreeBusyRequest object
}
\description{
FreeBusyRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other FreeBusyRequest functions: \code{\link{freebusy.query}}
}
