% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{FreeBusyRequestItem}
\alias{FreeBusyRequestItem}
\title{FreeBusyRequestItem Object}
\usage{
FreeBusyRequestItem(id = NULL)
}
\arguments{
\item{id}{The identifier of a calendar or a group}
}
\value{
FreeBusyRequestItem object
}
\description{
FreeBusyRequestItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

