% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{FreeBusyResponse.calendars}
\alias{FreeBusyResponse.calendars}
\title{FreeBusyResponse.calendars Object}
\usage{
FreeBusyResponse.calendars()
}
\value{
FreeBusyResponse.calendars object
}
\description{
FreeBusyResponse.calendars Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
List of free/busy information for calendars.
}
\seealso{
Other FreeBusyResponse functions: \code{\link{FreeBusyResponse.groups}},
  \code{\link{FreeBusyResponse}}
}
