% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{FreeBusyResponse.groups}
\alias{FreeBusyResponse.groups}
\title{FreeBusyResponse.groups Object}
\usage{
FreeBusyResponse.groups()
}
\value{
FreeBusyResponse.groups object
}
\description{
FreeBusyResponse.groups Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Expansion of groups.
}
\seealso{
Other FreeBusyResponse functions: \code{\link{FreeBusyResponse.calendars}},
  \code{\link{FreeBusyResponse}}
}
