% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{FreeBusyResponse}
\alias{FreeBusyResponse}
\title{FreeBusyResponse Object}
\usage{
FreeBusyResponse(FreeBusyResponse.calendars = NULL,
  FreeBusyResponse.groups = NULL, calendars = NULL, groups = NULL,
  timeMax = NULL, timeMin = NULL)
}
\arguments{
\item{FreeBusyResponse.calendars}{The \link{FreeBusyResponse.calendars} object or list of objects}

\item{FreeBusyResponse.groups}{The \link{FreeBusyResponse.groups} object or list of objects}

\item{calendars}{List of free/busy information for calendars}

\item{groups}{Expansion of groups}

\item{timeMax}{The end of the interval}

\item{timeMin}{The start of the interval}
}
\value{
FreeBusyResponse object
}
\description{
FreeBusyResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other FreeBusyResponse functions: \code{\link{FreeBusyResponse.calendars}},
  \code{\link{FreeBusyResponse.groups}}
}
