% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_functions.R
\name{ist.watch}
\alias{ist.watch}
\title{Watch for changes to CalendarList resources.}
\usage{
ist.watch(Channel, maxResults = NULL, minAccessRole = NULL,
  pageToken = NULL, showDeleted = NULL, showHidden = NULL,
  syncToken = NULL)
}
\arguments{
\item{Channel}{The \link{Channel} object to pass to this method}

\item{maxResults}{Maximum number of entries returned on one result page}

\item{minAccessRole}{The minimum access role for the user in the returned entries}

\item{pageToken}{Token specifying which result page to return}

\item{showDeleted}{Whether to include deleted calendar list entries in the result}

\item{showHidden}{Whether to show hidden entries}

\item{syncToken}{Token obtained from the nextSyncToken field returned on the last page of results from the previous list request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/calendar
\item https://www.googleapis.com/auth/calendar.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/calendar, https://www.googleapis.com/auth/calendar.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/google-apps/calendar/firstapp}{Google Documentation}

Other Channel functions: \code{\link{Channel.params}},
  \code{\link{Channel}}, \code{\link{acl.watch}},
  \code{\link{channels.stop}}, \code{\link{events.watch}},
  \code{\link{settings.watch}}
}
