% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{Setting}
\alias{Setting}
\title{Setting Object}
\usage{
Setting(etag = NULL, id = NULL, value = NULL)
}
\arguments{
\item{etag}{ETag of the resource}

\item{id}{The id of the user setting}

\item{value}{Value of the user setting}
}
\value{
Setting object
}
\description{
Setting Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

