% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{Settings}
\alias{Settings}
\title{Settings Object}
\usage{
Settings(etag = NULL, items = NULL, nextPageToken = NULL,
  nextSyncToken = NULL)
}
\arguments{
\item{etag}{Etag of the collection}

\item{items}{List of user settings}

\item{nextPageToken}{Token used to access the next page of this result}

\item{nextSyncToken}{Token used at a later point in time to retrieve only the entries that have changed since this result was returned}
}
\value{
Settings object
}
\description{
Settings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
