% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_objects.R
\name{TimePeriod}
\alias{TimePeriod}
\title{TimePeriod Object}
\usage{
TimePeriod(end = NULL, start = NULL)
}
\arguments{
\item{end}{The (exclusive) end of the time period}

\item{start}{The (inclusive) start of the time period}
}
\value{
TimePeriod object
}
\description{
TimePeriod Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
